modules/* moved to modules.py, added dockerignore, rebuilded dockerbuild, fixed little issues
