+++ .dockerignore
@@ -1,0 +1,5 @@
+*
+! models
+! main.py
+! pyproject.toml
+! uv.lock
