--- models/itemdb.py
@@ -1,5 +1,0 @@
-from models.items import Item
-
-
-class ItemsDB:
-    it: Item
