--- models/items.py
@@ -1,7 +1,0 @@
-from typing import List
-
-
-class Item:
-    coords: List = [int]
-    price: int = 0
-    date: int = 0
