+++ models/models.py
@@ -1,19 +1,90 @@
 import asyncio
 import logging
-import re
-from typing import List, Dict, Any
+from dataclasses import dataclass
 import requests
+import folium
+from folium.plugins import HeatMap
 
-from models.coords import Coords
-from models.requestedcoords import RequestedCoords
-from models.storage import Storage
 from string import Template
+import re
 
-from models.storageitems import StorageItem
+
+@dataclass
+class Coords:
+    x: float
+    y: float
+
+
+@dataclass
+class StorageItem:
+    coords: Coords
+    price: int
+
+
+class Storage:
+    __data: dict[int, StorageItem]
+
+    def __init__(self) -> None:
+        self.__data = {0: StorageItem(Coords(0, 0), 0)}
+
+    def update(self, updated_real_estates: dict[int, StorageItem]) -> None:
+        self.__data.update(updated_real_estates)
+
+    def return_with_filter(
+            self, x_max: float, x_min: float, y_max: float, y_min: float
+    ) -> dict[int, StorageItem]:
+        return {
+            key: value
+            for key, value in self.__data.items()
+            if float(x_min) < float(value.coords.x) < float(x_max) and float(y_min) < float(value.coords.y) < float(y_max) #todo fix some data is string, cant compare with float
+        }
+
+
+class RequestedCoords:
+    coords: list[Coords]
+
+    def __init__(self) -> None:
+        self.coords = []
+
+    def get_last(self) -> Coords | list[any]:
+        if self.count():
+            return self.coords[-1]
+        else:
+            return []
+
+    def delete_last(self) -> None:
+        if self.count():
+            self.coords.pop()
+
+    def count(self) -> int:
+        return len(self.coords)
+
+
+class ViewData:
+    file_path = "heatmap.html"
+
+    def __init__(self, storage: Storage) -> None:
+        self.storage = storage
+
+    def create_map(self, x: float, y: float) -> None:
+        city_center = [x, y]
+        data_points = []
+        logging.debug("Creating a map")
+        for val in self.storage.return_with_filter(
+                x + 0.02949858216, x - 0.02949858216, y + 0.05012512207, y - 0.05012512207
+        ).values():
+            data_points.append([val.coords.x, val.coords.y, val.price])
+
+        m = folium.Map(location=city_center, zoom_start=12)
+
+        heat_data = [[point[0], point[1], point[2]] for point in data_points]
+        HeatMap(heat_data).add_to(m)
+        m.save(self.file_path)
+        logging.debug("Map is created")
 
 
 class Scrapper:
-    coords_list: List = []
+    coords_list: list = []
     avito_url: str = (
         "https://www.avito.ru/js/1/map/items?categoryId=24&locationId=653240&correctorMode=0&page=$page"
         "&map=eyJzZWFyY2hBcmVhIjp7ImxhdEJvdHRvbSI6NTkuOTUxMzg3OTE4NjkwMDI0LCJsYXRUb3AiOjYwLjAxMDM4NzY0ODM"
@@ -61,7 +132,7 @@ class Scrapper:
 
     def __make_request(self, url: str) -> int:
 
-        data = Dict
+        data = dict
         total_count = int
         try:  # todo add 4xx 5xx handler and retry
             headers = {
@@ -77,7 +148,7 @@ class Scrapper:
         self.storage.update(self.__parce_info(data))
         return total_count
 
-    def __parce_info(self, raw_date: Dict[Any, Any]) -> Dict[int, StorageItem]:
+    def __parce_info(self, raw_date: dict[any, any]) -> dict[int, StorageItem]:
         real_state = {}
         for i in raw_date["items"]:
             real_state[int(i["id"])] = StorageItem(
